codes/futures-bot/requirements.txt
python-binance
python-dotenv
pandas
ta-lib
flask
flask-socketio